Changing the create user and create appointment forms

+++ petshopApp/src/main/kotlin/br/bridge/mariaeduarda/petshopApp/entities/User.kt
@@ -33,6 +33,6 @@ data class User(
         var name: String,
 
         @Column
-        var salary: Float
+        var salary: Int
 
 )
